requirements.txt
Flask==2.0.2
itsdangerous==2.1.2
Jinja2==3.0.3
MarkupSafe==2.1.3
pymongo==4.5.0
python-dotenv==1.0.0
PyYAML==6.0.1
Werkzeug==2.0.2
